主要要功能已实现

+++ fileOperator.h
@@ -2,6 +2,7 @@
 #define FILEOPERATOR_H
 #include <string>
 #include <fstream>
+#include <iostream>
 using namespace std;
 
 class fileOperator {
@@ -10,13 +11,15 @@ public:
     void merge(string json_file="config.json");
 private:
     bool exist(string name);
-    size_t file_size(ifstream &file);
-    void copy_file(ifstream &input, ofstream &output, size_t input_size);
+    long long file_size(ifstream &file);
+    void copy_file(ifstream &input, ofstream &output, long long input_size);
 
     inline void read_file_in_block(char* data, ifstream &input, int size=kBlockSize) {
+        //cout<<"read: "<<size<<endl;
         input.read(data, size);
     }
     inline void write_file_in_block(char* data, ofstream &output, int size=kBlockSize) {
+        //cout<<"write: "<<size<<endl;
         output.write(data, size);
     }
 private:
